% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{checkRemoteFileAvailable}
\alias{checkRemoteFileAvailable}
\title{Check if a given remote file is available for download}
\usage{
checkRemoteFileAvailable(remoteFile)
}
\arguments{
\item{remoteFile}{a remote resource}
}
\value{
NULL if the remote resource is not available, other "success"
}
\description{
Check if a given remote file is available for download
}
