% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patients.R
\name{checkTablesColumns}
\alias{checkTablesColumns}
\title{Check if the given tables and columns are valid and return the loaded data.}
\usage{
checkTablesColumns(cdmVersion, filePath, extraTable)
}
\arguments{
\item{cdmVersion}{cdm version}

\item{filePath}{Path to the test patient data in CSV format. Multiple CSV files representing tables tables from the OMOP-CDM must be provided, e.g. person.csv, drug_exposure.csv, condition_ocurrence.csv, etc.}

\item{extraTable}{extra table}
}
\value{
a named list containing the loaded data
}
\description{
Check if the given tables and columns are valid and return the loaded data.
}
