% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvertTestData.R
\name{convertJSONTestDataToSQL}
\alias{convertJSONTestDataToSQL}
\title{`convertJSONTestDataToSQL()` transforms patient json files to a SQL file.}
\usage{
convertJSONTestDataToSQL(pathToTestCases = NULL, pathToSqlFiles = NULL)
}
\arguments{
\item{pathToTestCases}{If NULL, takes the project path}

\item{pathToSqlFiles}{path where SQL file is stored. If NULL, it's using the pathToTestCases}
}
\value{
NULL, it writes a SQL file to disk
}
\description{
`convertJSONTestDataToSQL()` transforms patient json files to a SQL file.
}
