% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdmManagement.R
\name{emptyCDM}
\alias{emptyCDM}
\title{`emptyCDM()` truncates all the tables except those from the vocabulary for testing purposes}
\usage{
emptyCDM(con, cdm)
}
\arguments{
\item{con}{database connection}

\item{cdm}{cdm object}
}
\value{
A blank CDM with vocabulary
}
\description{
`emptyCDM()` truncates all the tables except those from the vocabulary for testing purposes
}
