% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisations.R
\name{graphCohort}
\alias{graphCohort}
\title{Deprecated cohort timeline visualisation}
\usage{
graphCohort(subject_id, cohorts = list())
}
\arguments{
\item{subject_id}{Deprecated.}

\item{cohorts}{Deprecated.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
\code{graphCohort()} is deprecated and will be removed in a future release.
}
