% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patients.R
\name{patientCDM}
\alias{patientCDM}
\title{`pushPatientCDM()` takes a file with patients in JSON format, pushes them into the blank CDM and performs the test.}
\usage{
patientCDM(filePathJson = NULL, testName = NULL)
}
\arguments{
\item{filePathJson}{If NULL, takes the project path to create the SQL files.}

\item{testName}{Name of the test patients files.}
}
\value{
Study results in the specified folder
}
\description{
`pushPatientCDM()` takes a file with patients in JSON format, pushes them into the blank CDM and performs the test.
}
