% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPatients.R
\name{patientSQL}
\alias{patientSQL}
\title{`testPatients()` takes a file with patients in JSON format, pushes them into the blank CDM and performs the test.}
\usage{
patientSQL(pathToTestCases = NULL)
}
\arguments{
\item{pathToTestCases}{If NULL, takes the project path to create the SQL files.}
}
\value{
Study results in the specified folder
}
\description{
`testPatients()` takes a file with patients in JSON format, pushes them into the blank CDM and performs the test.
}
