% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patients.R
\name{pushPatients}
\alias{pushPatients}
\title{`pushPatients()` pushes sample patients into a blank CDM if a connection is provided.}
\usage{
pushPatients(
  connection,
  cdmSchema,
  writeSchema,
  pathJson = NULL,
  testName = NULL
)
}
\arguments{
\item{connection}{A DBI connection to a blank CDM.}

\item{cdmSchema}{CDM schema of the database. E.g. "cdm" or "main".}

\item{writeSchema}{Writing schema of the database.}

\item{pathJson}{Directory where the sample populations in json are located. If NULL, gets the default inst/testCases directory.}

\item{testName}{Name of the sample population json file. If NULL it will push the first sample population in the testCases directory.}
}
\value{
A CDM reference object.
}
\description{
`pushPatients()` pushes sample patients into a blank CDM if a connection is provided.
}
