% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patients.R
\name{pushPatientSQL}
\alias{pushPatientSQL}
\title{`pushPatientSQL()` takes a file with patients in JSON format, pushes them into the blank CDM and performs the test.}
\usage{
pushPatientSQL(pathJson = NULL)
}
\arguments{
\item{pathJson}{If NULL, takes the project path to create the SQL files}
}
\value{
Study results in the specified folder
}
\description{
`pushPatientSQL()` takes a file with patients in JSON format, pushes them into the blank CDM and performs the test.
}
