% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patients.R
\name{sqlPatients}
\alias{sqlPatients}
\title{`sqlPatients()` takes a file with patients in JSON format, pushes them into the blank CDM using DatabaseConnector.}
\usage{
sqlPatients(pathJson = NULL, testName = NULL)
}
\arguments{
\item{pathJson}{If NULL, takes the project path to create the SQL files}
}
\value{
Study results in the specified folder
}
\description{
`sqlPatients()` takes a file with patients in JSON format, pushes them into the blank CDM using DatabaseConnector.
}
