% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teststudy.R
\name{testPatients}
\alias{testPatients}
\title{`testPatients()` takes a file with patients in SQL format, pushes them into the blank CDM and performs the test.}
\usage{
testPatients(
  dbConnection,
  functionsToTest = NULL,
  createSchemaPerTest = FALSE,
  cdmDatabaseSchema = Sys.getenv("UT_CDM_SCHEMA"),
  cohortDatabaseSchema = Sys.getenv("UT_COHORT_SCHEMA"),
  unitTestOutputFolder = Sys.getenv("UT_TEST_CASES_RESULTS_LOCATION"),
  user = Sys.getenv("UT_DB_USER"),
  password = Sys.getenv("UT_DB_PASSWORD"),
  dbms = Sys.getenv("UT_DBMS"),
  dbname = Sys.getenv("UT_DB_NAME"),
  server = Sys.getenv("UT_DB_SERVER"),
  port = Sys.getenv("UT_DB_PORT"),
  cdm_version = "5.4",
  outputFolder = "results"
)
}
\arguments{
\item{dbConnection}{database connection}

\item{functionsToTest}{the functions to test}

\item{createSchemaPerTest}{if a schema should be created per test}

\item{cdmDatabaseSchema}{cdm schema}

\item{cohortDatabaseSchema}{cohort schema}

\item{unitTestOutputFolder}{unit test output folder}

\item{user}{db username}

\item{password}{db pwd}

\item{dbms}{dbms name}

\item{dbname}{database name}

\item{server}{database server}

\item{port}{database port number}

\item{cdm_version}{cdm version number}

\item{outputFolder}{output folder}
}
\value{
Study results in the specified folder
}
\description{
`testPatients()` takes a file with patients in SQL format, pushes them into the blank CDM and performs the test.
}
