% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patients.R
\name{checkTablesColumns}
\alias{checkTablesColumns}
\title{Check if the given tables and columns are valid and return the loaded data.}
\usage{
checkTablesColumns(cdmVersion, filePath, extraTable = FALSE)
}
\arguments{
\item{cdmVersion}{cdm version}

\item{filePath}{Path to the test patient data in xlsx format}

\item{extraTable}{if extra tables are provided or not, default FALSE}
}
\value{
a named list containing the loaded data
}
\description{
Check if the given tables and columns are valid and return the loaded data.
}
