% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilscdm.R
\name{emptyCDM}
\alias{emptyCDM}
\title{`emptyCDM()` truncates all the tables except those from the vocabulary for testing purposes}
\usage{
emptyCDM(conn, cdm)
}
\value{
A blanck CDM with vocabulary
}
\description{
`emptyCDM()` truncates all the tables except those from the vocabulary for testing purposes
}
