% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateTestTable.R
\name{generateTestTables}
\alias{generateTestTables}
\title{Generates an Excel file with sheets that correspond to an OMOP-CDM tables.}
\usage{
generateTestTables(tableNames, cdmVersion, outputFolder)
}
\arguments{
\item{tableNames}{A list specifying the table names to include in the Excel file.}

\item{outputFolder}{The folder where the Excel file will be saved.}

\item{cdmversion}{The CDM version to use for creating the requested tables (either 5.3 or 5.4).}
}
\value{
An Excel file with the tables requested.
}
\description{
Generates an Excel file with sheets that correspond to an OMOP-CDM tables.
}
