% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patients.R
\name{patientsCDM}
\alias{patientsCDM}
\title{`patientCDM()` takes a file with patients in JSON format, pushes them into the blank CDM and performs the test.}
\usage{
patientsCDM(pathJson = proj_path("inst", "testCases"), testName = NULL)
}
\arguments{
\item{pathJson}{If NULL, takes the project path to create the SQL files.}

\item{testName}{Name of the Unit Test Definition, if NULL it will push the first sample population in the testCases directory.}
}
\value{
Study results in the specified folder
}
\description{
`patientCDM()` takes a file with patients in JSON format, pushes them into the blank CDM and performs the test.
}
