% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patients.R
\name{readPatients}
\alias{readPatients}
\title{`readPatients()` converts a test patients in XLSX format into a JSON for testing, and creates a JSON in the a inst/testCases folder.}
\usage{
readPatients(
  filePath = NULL,
  testName = "test",
  sheets = c("person", "observation_period", "drug_exposure", "condition_occurrence",
    "visit_occurrence", "visit_context", "visit_detail", "death"),
  outputPath = NULL
)
}
\arguments{
\item{filePath}{Path to the test patient data in Excel format.}

\item{testName}{Name of the test population in character.}

\item{sheets}{List of sheets to be converted into tables.}

\item{outputPath}{Path of the output file. If NULL}
}
\value{
A JSON file for testing inside the package directory of a DARWIN EU study.
}
\description{
`readPatients()` converts a test patients in XLSX format into a JSON for testing, and creates a JSON in the a inst/testCases folder.
}
