% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patients.R
\name{readPatients}
\alias{readPatients}
\title{`readPatients()` converts a sample of patients in XLSX format into Unit Testing Definition in JSON format.}
\usage{
readPatients(filePath = NULL, testName = "test", outputPath = NULL)
}
\arguments{
\item{filePath}{Path to the test patient data in Excel format.}

\item{testName}{Name of the test population in character.}

\item{outputPath}{Path of the output file, iff NULL, a folder will be created in the project folder inst/testCases}
}
\value{
A JSON file for testing inside the project directory.
}
\description{
`readPatients()` converts a sample of patients in XLSX format into Unit Testing Definition in JSON format.
}
