% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPatients.R
\name{readPatients}
\alias{readPatients}
\title{`readPatients()` converts a test patients into a SQL file for testing.}
\usage{
readPatients(filePath = NULL)
}
\arguments{
\item{dataPath}{Path to the test patient data in Excel format.}
}
\value{
A SQL file for testing
}
\description{
`readPatients()` converts a test patients into a SQL file for testing.
}
