% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testPatients.R
\name{testPatients}
\alias{testPatients}
\title{`testPatients()` takes a file with patients in JSON format, pushes them into the blank CDM and performs the test.}
\usage{
testPatients(
  dbConnection,
  functionsToTest = NULL,
  createSchemaPerTest = FALSE,
  cdmDatabaseSchema = Sys.getenv("UT_CDM_SCHEMA"),
  cohortDatabaseSchema = Sys.getenv("UT_COHORT_SCHEMA"),
  unitTestOutputFolder = Sys.getenv("UT_TEST_CASES_RESULTS_LOCATION"),
  user = Sys.getenv("UT_DB_USER"),
  password = Sys.getenv("UT_DB_PASSWORD"),
  dbms = Sys.getenv("UT_DBMS"),
  dbname = Sys.getenv("UT_DB_NAME"),
  server = Sys.getenv("UT_DB_SERVER"),
  port = Sys.getenv("UT_DB_PORT"),
  cdm_version = "5.4",
  outputFolder = here::here("results")
)
}
\arguments{
\item{testCaseFile}{Path to JSON test files. Those should contain patients, observation period, drug exposure, condition and visit occurrence.}
}
\value{
Study results in the specified folder
}
\description{
`testPatients()` takes a file with patients in JSON format, pushes them into the blank CDM and performs the test.
}
